Move the typing tests to the folders of the modules they concern
