+++ handsontable/src/__tests__/core/base.types.ts
