+++ handsontable/src/__tests__/core/base.unit.js
@@ -1,11 +1,11 @@
 // eslint-disable-next-line no-unused-vars
 import Handsontable from 'handsontable/base';
-import { HOT_PACKAGE_NAME, HOT_VERSION } from '../../../hot.config.js';
-import { getRegisteredCellTypeNames } from '../cellTypes/registry';
-import { getRegisteredEditorNames } from '../editors/registry';
-import { getPluginsNames } from '../plugins/registry';
-import { getRegisteredRendererNames } from '../renderers/registry';
-import { getRegisteredValidatorNames } from '../validators/registry';
+import { HOT_PACKAGE_NAME, HOT_VERSION } from '../../../../hot.config.js';
+import { getRegisteredCellTypeNames } from '../../cellTypes/registry';
+import { getRegisteredEditorNames } from '../../editors/registry';
+import { getPluginsNames } from '../../plugins/registry';
+import { getRegisteredRendererNames } from '../../renderers/registry';
+import { getRegisteredValidatorNames } from '../../validators/registry';
 
 describe('`handsontable/base` entry point', () => {
   it('should by default contain only text cell meta module', () => {
