+++ handsontable/src/__tests__/core/cellCoords.types.ts
