+++ handsontable/src/__tests__/core/cellRange.types.ts
