+++ handsontable/src/__tests__/core/core.types.ts
