+++ handsontable/src/__tests__/core/methods.types.ts
