+++ handsontable/src/__tests__/core/settings.types.ts
