+++ handsontable/src/editors/__tests__/editors.types.ts
