+++ handsontable/src/editors/passwordEditor/__tests__/password.types.ts
