+++ handsontable/src/editors/textEditor/__tests__/text.types.ts
