+++ handsontable/src/helpers/__tests__/helpers.types.ts
