+++ handsontable/src/helpers/dom/__tests__/dom.types.ts
