+++ handsontable/src/i18n/__tests__/i18n.types.ts
