+++ handsontable/src/plugins/__tests__/plugins.types.ts
