+++ handsontable/src/plugins/__tests__/registry.types.ts
