+++ handsontable/src/renderers/__tests__/renderers.types.ts
