+++ handsontable/src/renderers/__tests__/renderersModule.types.ts
