+++ handsontable/src/validators/__tests__/validators.types.ts
